requirements.txt
Django>=5.1,<7.0
psycopg2-binary>=2.9
django-htmx>=1.21
django-imagekit>=5.0
django-modeltranslation>=0.19
whitenoise>=6.8
gunicorn>=23.0
Pillow>=11.0
requests>=2.32
beautifulsoup4>=4.12
python-dotenv>=1.0
